requirements.txt
Django>=4.0.1,<4.1
djangorestframework==3.14.0
mysqlclient==2.1.1
mysql-connector
djangorestframework-simplejwt
#celery
celery
celery-beat
# Redis support
redis
django-redis==5.2.0